refactored appServer so that the server imports the routes for cats whenever the cats API is called, restructured file system to account for this change in modularity
