+++ lib/routes/deleteHandler.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const resourceDir = './resources/';
+const resourceDir = './lib/models/resources/';
 
 function deleteHandler(reqParamsObject, deleteResponse) {
     var reqUrl = `${resourceDir}${reqParamsObject['id']}.json`;
