+++ lib/routes/getHandler.js
@@ -1,5 +1,5 @@
 const sander = require('sander');
-const resourceDir = './resources';
+const resourceDir = './lib/models/resources/';
 
 function getHandler(reqParamsObject, getResponse) {
     var firstArg = reqParamsObject['0'];
