+++ lib/routes/postHandler.js
@@ -1,5 +1,5 @@
 const sander = require('sander');
-const resourceDir = './resources/';
+const resourceDir = './lib/models/resources/';
 
 function postHandler(reqParamsObject, req, postResponse) {
     var body = '';
