+++ lib/routes/putHandler.js
@@ -1,5 +1,5 @@
 const sander = require('sander');
-const resourceDir = './resources/';
+const resourceDir = './lib/models/resources/';
 
 function updateJson(originalJsonData, updatedJsonData) {
     var jsonKeys = Object.keys(updatedJsonData);
