+++ package.json
@@ -23,7 +23,8 @@
     "chai": "^3.5.0",
     "chai-http": "^3.0.0",
     "eslint": "^3.9.1",
-    "mocha": "^3.1.2"
+    "mocha": "^3.1.2",
+    "morgan": "^1.7.0"
   },
   "dependencies": {
     "express": "^4.14.0",
